word is ugly bxtch
This is the very first edition of the explaination of the whole data pre-process work. read me if I run from this project
btw, I wrote this in notion, but I have to copy this to a word file.
WORD is a ugly bxtch, why don't everyone just use notion.